refactor(app): drop settings facade
Move callers onto the explicit config owners and fail if the\ncompatibility facade path returns.\n\nCloses #2919\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
