+++ CONTRIBUTING.md
@@ -141,7 +141,7 @@ ownership boundaries only when that distinction matters to maintenance.
 
 The backend uses a layered config system. Values are merged in this order (later wins):
 
-1. **Built-in defaults** (`vibesensor.app.settings.DEFAULT_CONFIG`) — always present, never edited
+1. **Built-in defaults** (`vibesensor.app.config_defaults.DEFAULT_CONFIG`) — always present, never edited
 2. **`config.yaml`** — local overrides (gitignored; mostly empty by default)
 3. **Environment variables** — override individual keys at runtime
