+++ apps/server/vibesensor/cli/preflight.py
@@ -4,10 +4,10 @@ import argparse
 import json
 from pathlib import Path
 
-from vibesensor.app.settings import (
-    AppConfig,
-    documented_default_config,
-    load_config,
+from vibesensor.app.config_defaults import documented_default_config
+from vibesensor.app.config_loader import load_config
+from vibesensor.app.config_schema import AppConfig
+from vibesensor.shared.process_settings import (
     summarize_process_settings,
 )
 from vibesensor.shared.constants.ui import UI_HEAVY_PUSH_HZ, UI_PUSH_HZ
